refactor to UserCreateEvent
