+++ src/system/UsersModule/Event/DeletedRegistrationEvent.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Zikula\UsersModule\Event;
 
-use Zikula\UsersModule\Entity\UserEntity;
-
 /**
  * Occurs after a registration record is deleted. This could occur as a result of the administrator deleting the
  * record through the approval/denial process, or it could happen because the registration request expired. This
@@ -23,31 +21,6 @@ use Zikula\UsersModule\Entity\UserEntity;
  * UI-level actions such as redirects.
  * The subject of the event is set to the Uid being deleted.
  */
-class DeletedRegistrationEvent
+class DeletedRegistrationEvent extends UserEntityEvent
 {
-    /**
-     * @var UserEntity
-     */
-    private $user;
-
-    /**
-     * @var \DateTimeImmutable
-     */
-    private $date;
-
-    public function __construct(UserEntity $user)
-    {
-        $this->user = $user;
-        $this->date = new \DateTimeImmutable('now');
-    }
-
-    public function getUser(): UserEntity
-    {
-        return $this->user;
-    }
-
-    public function getDate(): \DateTimeImmutable
-    {
-        return $this->date;
-    }
 }
