+++ src/system/UsersModule/Event/StoppableUserEntityEvent.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Zikula package.
+ *
+ * Copyright Zikula Foundation - https://ziku.la/
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Zikula\UsersModule\Event;
+
+use Psr\EventDispatcher\StoppableEventInterface;
+
+/**
+ * A 'Generic' event that requires a UserEntity on construction and sets
+ * an immutable datetime object for tracking purposes.
+ * This class adds the ability to stop the propagation.
+ */
+class StoppableUserEntityEvent extends UserEntityEvent implements StoppableEventInterface
+{
+    private $propagationStopped = false;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isPropagationStopped(): bool
+    {
+        return $this->propagationStopped;
+    }
+
+    /**
+     * Stops the propagation of the event to further event listeners.
+     *
+     * If multiple event listeners are connected to the same event, no
+     * further event listener will be triggered once any trigger calls
+     * stopPropagation().
+     */
+    public function stopPropagation(): void
+    {
+        $this->propagationStopped = true;
+    }
+}
