+++ src/system/UsersModule/RegistrationEvents.php
@@ -25,27 +25,6 @@ class RegistrationEvents
     public const REGISTRATION_STARTED = 'module.users.ui.registration.started';
 
     /**
-     * Occurs when the Registration process is determining whether to create a 'registration' or a 'full user'.
-     *
-     * The subject of the event is the UserEntity. There are no arguments or data. If the User hasn't been persisted, then
-     * there will be no Uid.
-     *
-     * A handler that needs to veto a registration should call `stopPropagation()`. This will prevent other handlers
-     * from receiving the event, will return to the registration process, and will prevent the registration from
-     * creating a 'full user' record.
-     *
-     * For example an authentication method may veto a registration attempt if it requires a user to verify some
-     * registration data by email.
-     *
-     * It is assumed that the authentication method will have notified the user of required steps to prevent future
-     * vetoes. And provide the methods to correct the issue and process the steps.
-     *
-     * Because this event will not necessarily notify ALL listeners (if propagation is stopped) it CANNOT be relied upon
-     * to effect change of any kind with regard to the entity.
-     */
-    public const FULL_USER_CREATE_VETO = 'full.user.create.veto';
-
-    /**
      * Occurs after a user has successfully registered a new account in the system. It will follow either a
      * `registration.create` event, or a `user.create` event, depending on the result of the registration process, the
      * information provided by the user, and several configuration options set in the Users module. The resultant record
